services/embedding/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
PyMuPDF
python-docx
python-pptx
python-multipart
aiofiles